Add data files via upload
dataset.dat : The data set used in this project. It is the union of the train set and the test set. 
duration.dat : The duration information extracted from the dataset.dat file.
pitch.dat : The pitch information extracted from the dataset.dat file.
(pitch.dat is divided into two parts, for its size is over 25 Mbytes, which may be the limited size of the fie uploaded...I suppose)